requirements.txt
openai-agents>=0.13.0
scikit-learn>=1.8.0
pandas>=2.2.0
numpy>=1.26.0
fastapi>=0.130.0
uvicorn>=0.40.0
pydantic>=2.0.0
python-dotenv>=1.0.0
joblib>=1.3.0
httpx>=0.27.0
